requirements.txt
selenium
python-whois
alphabet-detector